+++ container-engine-lib/lib/backend_impls/docker/docker_kurtosis_backend_user_service_functions.go
@@ -1,32 +1,32 @@
 package docker
-
 import (
 	"bytes"
 	"context"
 	"fmt"
+	"github.com/docker/docker/pkg/stdcopy"
 	"github.com/docker/go-connections/nat"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_log_streaming_readcloser"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_manager"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_manager/types"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/docker_operation_parallelizer"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/object_attributes_provider"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/object_attributes_provider/docker_port_spec_serializer"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/object_attributes_provider/label_key_consts"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_impls/docker/object_attributes_provider/label_value_consts"
+	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/container_status"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/enclave"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/exec_result"
-	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/files_artifact_expansion"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/port_spec"
 	"github.com/kurtosis-tech/container-engine-lib/lib/backend_interface/objects/service"
+	"github.com/kurtosis-tech/container-engine-lib/lib/concurrent_writer"
+	"github.com/kurtosis-tech/free-ip-addr-tracker-lib/lib"
 	"github.com/kurtosis-tech/stacktrace"
 	"github.com/sirupsen/logrus"
 	"io"
 	"net"
+	"strings"
 	"time"
-)
-
-const (
-	shouldGetStoppedContainersWhenGettingServiceInfo = true
 )
 
 /*
@@ -75,6 +75,15 @@ The benefits of this implementation:
 - We can call the GetServices method (that the CLI needs) without the API container running
  */
 
+const (
+	shouldGetStoppedContainersWhenGettingServiceInfo = true
+
+	shouldFollowContainerLogsWhenExpanderHasError = false
+
+	expanderContainerSuccessExitCode = 0
+)
+
+
 // We'll try to use the nicer-to-use shells first before we drop down to the lower shells
 var commandToRunWhenCreatingUserServiceShell = []string{
 	"sh",
@@ -83,9 +92,15 @@ var commandToRunWhenCreatingUserServiceShell = []string{
 }
 
 
+// NOTE: Normally we'd have a "canonical" resource here, where that resource is always guaranteed to exist. For Kurtosis services,
+// we want this to be the container engine's representation of a user service registration. Unfortunately, Docker has no way
+// of representing a user service registration, so we store them in an in-memory map on the DockerKurtosisBackend. Therefore, an
+// entry in that map is actually the canonical representation, which means that any of these fields could be nil!
 type userServiceDockerResources struct {
-	// This will never be nil because a user services is represented ONLY by a container in Docker
-	container *types.Container
+	serviceContainer *types.Container
+
+	// Will never be nil but may be empty if no expander volumes exist
+	expanderVolumeNames []string
 }
 
 func (backend *DockerKurtosisBackend) RegisterUserService(ctx context.Context, enclaveId enclave.EnclaveID, serviceId service.ServiceID, ) (*service.ServiceRegistration, error, ) {
@@ -149,9 +164,37 @@ func (backend *DockerKurtosisBackend) RegisterUserService(ctx context.Context, e
 	return registration, nil
 }
 
-func (backend *DockerKurtosisBackend) StartUserService(ctx context.Context, enclaveId enclave.EnclaveID, serviceGuid service.ServiceGUID, containerImageName string, privatePorts map[string]*port_spec.PortSpec, entrypointArgs []string, cmdArgs []string, envVars map[string]string, filesArtifactVolumeMountDirpaths map[files_artifact_expansion.FilesArtifactExpansionGUID]string, ) (*service.Service, error, ) {
+func (backend *DockerKurtosisBackend) StartUserService(
+	ctx context.Context,
+	enclaveId enclave.EnclaveID,
+	serviceGuid service.ServiceGUID,
+	containerImageName string,
+	privatePorts map[string]*port_spec.PortSpec,
+	entrypointArgs []string,
+	cmdArgs []string,
+	envVars map[string]string,
+	filesArtifactsExpansion *backend_interface.FilesArtifactsExpansion,
+) (*service.Service, error, ) {
 	backend.serviceRegistrationMutex.Lock()
 	defer backend.serviceRegistrationMutex.Unlock()
+
+	enclaveNetwork, err := backend.getEnclaveNetworkByEnclaveId(ctx, enclaveId)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred getting enclave network by enclave ID '%v'", enclaveId)
+	}
+	enclaveNetworkId := enclaveNetwork.GetId()
+
+	// Needed for files artifacts expansion container
+	freeIpAddrProvider, found := backend.enclaveFreeIpProviders[enclaveId]
+	if !found {
+		return nil, stacktrace.NewError(
+			"Received a request to start service with GUID '%v' in enclave '%v', but no free IP address provider was "+
+				"defined for this enclave; this likely means that the start request is being called where it shouldn't "+
+				"be (i.e. outside the API container)",
+			serviceGuid,
+			enclaveId,
+		)
+	}
 
 	// Find the registration
 	registrationsForEnclave, found := backend.serviceRegistrations[enclaveId]
@@ -191,6 +234,39 @@ func (backend *DockerKurtosisBackend) StartUserService(ctx context.Context, encl
 		return nil, stacktrace.Propagate(err, "Couldn't get an object attribute provider for enclave '%v'", enclaveId)
 	}
 
+	volumeMounts := map[string]string{}
+	shouldDeleteVolumes := true
+	if filesArtifactsExpansion != nil {
+		candidateVolumeMounts, err := backend.doFilesArtifactExpansionAndGetUserServiceVolumes(
+			ctx,
+			serviceGuid,
+			enclaveObjAttrsProvider,
+			freeIpAddrProvider,
+			enclaveNetworkId,
+			filesArtifactsExpansion.ExpanderImage,
+			filesArtifactsExpansion.ExpanderEnvVars,
+			filesArtifactsExpansion.ExpanderDirpathsToServiceDirpaths,
+		)
+		if err != nil {
+			return nil, stacktrace.Propagate(
+				err,
+				"An error occurred doing files artifacts expansion to get user service volumes",
+			)
+		}
+		defer func() {
+			if shouldDeleteVolumes {
+				for volumeName := range candidateVolumeMounts {
+					// Use background context so we delete these even if input context was cancelled
+					if err := backend.dockerManager.RemoveVolume(context.Background(), volumeName); err != nil {
+						logrus.Errorf("Starting the service failed so we tried to delete files artifact expansion volume '%v' that we created, but doing so threw an error:\n%v", volumeName, err)
+						logrus.Errorf("You'll need to delete volume '%v' manually!", volumeName)
+					}
+				}
+			}
+		}()
+		volumeMounts = candidateVolumeMounts
+	}
+
 	containerAttrs, err := enclaveObjAttrsProvider.ForUserServiceContainer(
 		serviceId,
 		serviceGuid,
@@ -207,11 +283,6 @@ func (backend *DockerKurtosisBackend) StartUserService(ctx context.Context, encl
 		labelStrs[labelKey.GetString()] = labelValue.GetString()
 	}
 
-	enclaveNetwork, err := backend.getEnclaveNetworkByEnclaveId(ctx, enclaveId)
-	if err != nil {
-		return nil, stacktrace.Propagate(err, "An error occurred getting enclave network by enclave ID '%v'", enclaveId)
-	}
-
 	dockerUsedPorts := map[nat.Port]docker_manager.PortPublishSpec{}
 	for portId, portSpec := range privatePorts {
 		dockerPort, err := transformPortSpecToDockerPort(portSpec)
@@ -224,7 +295,7 @@ func (backend *DockerKurtosisBackend) StartUserService(ctx context.Context, encl
 	createAndStartArgsBuilder := docker_manager.NewCreateAndStartContainerArgsBuilder(
 		containerImageName,
 		containerName.GetString(),
-		enclaveNetwork.GetId(),
+		enclaveNetworkId,
 	).WithStaticIP(
 		privateIpAddr,
 	).WithUsedPorts(
@@ -242,37 +313,10 @@ func (backend *DockerKurtosisBackend) StartUserService(ctx context.Context, encl
 	if cmdArgs != nil {
 		createAndStartArgsBuilder.WithCmdArgs(cmdArgs)
 	}
-
-	if filesArtifactVolumeMountDirpaths != nil {
-
-		// CREATE FILTER SET FOR FILES ARTIFACT EXPANSION GUIDS AND SERVICE GUID
-		filterFilesArtifactExpansionGuids := map[files_artifact_expansion.FilesArtifactExpansionGUID]bool{}
-		for guid, _ := range filesArtifactVolumeMountDirpaths {
-			filterFilesArtifactExpansionGuids[guid] = true
-		}
-		filters := files_artifact_expansion.FilesArtifactExpansionFilters{
-			GUIDs: filterFilesArtifactExpansionGuids,
-			ServiceGUIDs: map[service.ServiceGUID]bool{
-				serviceGuid: true,
-			},
-		}
-		// GET MATCHING EXPANSION OBJECTS AND RESOURCES FOR FILTERS
-		filteredObjectsAndResources, err := backend.getMatchingFilesArtifactExpansionObjectsAndDockerResources(ctx, enclaveId, &filters)
-		if err != nil {
-			return nil, stacktrace.Propagate(err, "Failed to get expansion volumes for filters '%+v'", filters)
-		}
-		// ITERATE THROUGH MATCHING OBJECTS AND PLACE VOLUME NAMES FROM DOCKER OBJECTS INTO MAP FOR VOLUME TO PATH MAPPING
-		filesArtifactVolumeMountDirpathStrs := map[string]string{}
-		for filesArtifactGUID, objectsAndResources := range filteredObjectsAndResources {
-			resources := objectsAndResources.dockerResources
-			if resources == nil {
-				return nil, stacktrace.Propagate(err, "Found expansion with no docker resources for files artifact expansion '%v'", filesArtifactGUID)
-			}
-			volume := resources.volume
-			filesArtifactVolumeMountDirpathStrs[volume.Name] = filesArtifactVolumeMountDirpaths[filesArtifactGUID]
-		}
-		createAndStartArgsBuilder.WithVolumeMounts(filesArtifactVolumeMountDirpathStrs)
+	if volumeMounts != nil {
+		createAndStartArgsBuilder.WithVolumeMounts(volumeMounts)
 	}
+
 	createAndStartArgs := createAndStartArgsBuilder.Build()
 
 	// Best-effort pull attempt
@@ -319,6 +363,7 @@ func (backend *DockerKurtosisBackend) StartUserService(ctx context.Context, encl
 		maybePublicPortSpecs,
 	)
 
+	shouldDeleteVolumes = false
 	shouldKillContainer = false
 	return result, nil
 }
@@ -358,7 +403,7 @@ func (backend *DockerKurtosisBackend) GetUserServiceLogs(
 	erroredUserServices := map[service.ServiceGUID]error{}
 	shouldCloseLogStreams := true
 	for guid, resourcesForService := range allDockerResources {
-		container := resourcesForService.container
+		container := resourcesForService.serviceContainer
 		if container == nil {
 			erroredUserServices[guid] = stacktrace.NewError("Cannot get logs for service '%v' as it has no container", guid)
 			continue
@@ -399,7 +444,7 @@ func (backend *DockerKurtosisBackend) PauseService(
 	if err != nil {
 		return stacktrace.Propagate(err, "Failed to get information about service '%v' from Kurtosis backend.", serviceGuid)
 	}
-	container := dockerResources.container
+	container := dockerResources.serviceContainer
 	if container == nil {
 		return stacktrace.NewError("Cannot pause service '%v' as it doesn't have a container to pause", serviceGuid)
 	}
@@ -418,7 +463,7 @@ func (backend *DockerKurtosisBackend) UnpauseService(
 	if err != nil {
 		return stacktrace.Propagate(err, "Failed to get information about service '%v' from Kurtosis backend.", serviceGuid)
 	}
-	container := dockerResources.container
+	container := dockerResources.serviceContainer
 	if container == nil {
 		return stacktrace.NewError("Cannot unpause service '%v' as it doesn't have a container to pause", serviceGuid)
 	}
@@ -465,7 +510,7 @@ func (backend *DockerKurtosisBackend) RunUserServiceExecCommands(
 			)
 			continue
 		}
-		container := dockerResources.container
+		container := dockerResources.serviceContainer
 
 		execOutputBuf := &bytes.Buffer{}
 		exitCode, err := backend.dockerManager.RunExecCommand(
@@ -504,7 +549,7 @@ func (backend *DockerKurtosisBackend) GetConnectionWithUserService(
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred getting service object and Docker resources for service '%v' in enclave '%v'", serviceGuid, enclaveId)
 	}
-	container := serviceDockerResources.container
+	container := serviceDockerResources.serviceContainer
 
 	hijackedResponse, err := backend.dockerManager.CreateContainerExec(ctx, container.GetId(), commandToRunWhenCreatingUserServiceShell)
 	if err != nil {
@@ -530,7 +575,7 @@ func (backend *DockerKurtosisBackend) CopyFilesFromUserService(
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred getting user service with GUID '%v' in enclave with ID '%v'", serviceGuid, enclaveId)
 	}
-	container := serviceDockerResources.container
+	container := serviceDockerResources.serviceContainer
 
 	tarStreamReadCloser, err := backend.dockerManager.CopyFromContainer(ctx, container.GetId(), srcPathOnContainer)
 	if err != nil {
@@ -578,7 +623,7 @@ func (backend *DockerKurtosisBackend) StopUserServices(
 			// Should never happen; there should be a 1:1 mapping between service_objects:docker_resources by GUID
 			return nil, nil, stacktrace.NewError("No service object found for service '%v' that had Docker resources", guid)
 		}
-		servicesToStopByContainerId[serviceResources.container.GetId()] = serviceObj
+		servicesToStopByContainerId[serviceResources.serviceContainer.GetId()] = serviceObj
 	}
 
 	// TODO PLEAAASE GO GENERICS... but we can't use 1.18 yet because it'll break all Kurtosis clients :(
@@ -621,6 +666,16 @@ func (backend *DockerKurtosisBackend) StopUserServices(
 	return successfulGuids, erroredGuids, nil
 }
 
+/*
+!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+This code is INCREDIBLY tricky, as a result of:
+1) Needing to do service registrations to get an IP address before the service container is started
+2) Docker not having a canonical way to represent a service registration-before-container-started,
+   which requires us to use an in-memory registration map
+
+        Be VERY careful when modifying this code, and ideally get Kevin's eyes on it!!
+!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! WARNING !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
+ */
 func (backend *DockerKurtosisBackend) DestroyUserServices(
 	ctx context.Context,
 	enclaveId enclave.EnclaveID,
@@ -652,6 +707,8 @@ func (backend *DockerKurtosisBackend) DestroyUserServices(
 		)
 	}
 
+	// We filter registrations here because a registration is the canonical resource for a Kurtosis user service - no registration,
+	// no Kurtosis service - and not all registrations will have Docker resources
 	matchingRegistrations := map[service.ServiceGUID]*service.ServiceRegistration{}
 	for guid, registration := range registrationsForEnclave {
 		if filters.GUIDs != nil && len(filters.GUIDs) > 0 {
@@ -669,75 +726,76 @@ func (backend *DockerKurtosisBackend) DestroyUserServices(
 		matchingRegistrations[guid] = registration
 	}
 
+	// NOTE: This may end up with less results here than we have registrations, if the user registered but did not start a service,
+	// though we should never end up with _more_ Docker resources
 	allServiceObjs, allDockerResources, err := backend.getMatchingUserServiceObjsAndDockerResourcesNoMutex(ctx, enclaveId, filters)
 	if err != nil {
 		return nil, nil, stacktrace.Propagate(err, "An error occurred getting user services matching filters '%+v'", filters)
 	}
 
-	// Sanity check
-	for guid := range allDockerResources {
-		if _, found := matchingRegistrations[guid]; !found {
-			// Should never happen
-			return nil, nil, stacktrace.NewError("Service '%v' has Docker resources but no container registration; this is a bug in Kurtosis", guid)
+	if len(allServiceObjs) > len(matchingRegistrations) || len(allDockerResources) > len(matchingRegistrations) {
+		return nil, nil, stacktrace.NewError("Found more Docker resources matching the following filters than service registrations; this is a bug in Kurtosis: %+v", filters)
+	}
+
+	// TODO Refactor getMatchingUserServiceObjsAndDockerResourcesNoMutex to return a single package object so we don't need to do this
+	// Before deleting anything, verify we have the same keys in both maps
+	if len(allServiceObjs) != len(allDockerResources) {
+		return nil, nil, stacktrace.NewError(
+			"Expected the number of services to remove (%v), and the number of services for which we have Docker resources for (%v), to be the same but were different!",
+			len(allServiceObjs),
+			len(allDockerResources),
+		)
+	}
+	for serviceGuid := range allServiceObjs {
+		if _, found := allDockerResources[serviceGuid]; !found {
+			return nil, nil, stacktrace.NewError(
+				"Have service object to remove '%v', which doesn't have corresponding Docker resources; this is a " +
+					"bug in Kurtosis",
+				serviceGuid,
+			)
 		}
 	}
 
 	registrationsToDeregister := map[service.ServiceGUID]*service.ServiceRegistration{}
-	servicesToDestroyByContainerIdBeforeDeregistration := map[string]interface{}{}
+
+	// Find the registrations which don't have any Docker resources and immediately add them to the list of stuff to deregister
 	for guid, registration := range matchingRegistrations {
-		dockerResources, found := allDockerResources[guid]
-		if !found {
-			// For registrations-without-containers, only add them to the deregistration list if the status filter wasn't specified
-			if filters.Statuses == nil || len(filters.Statuses) == 0 {
-				registrationsToDeregister[guid] = registration
-			}
+		if _, doesRegistrationHaveResources := allDockerResources[guid]; doesRegistrationHaveResources {
+			// We'll deregister registrations-with-resources if and only if we can successfully remove their resources
 			continue
 		}
-		containerId := dockerResources.container.GetId()
 
-		serviceObj, found := allServiceObjs[guid]
-		if !found {
-			// Should never happen
-			return nil, nil, stacktrace.NewError("Service '%v' has Docker resources but no service object; this is a bug in Kurtosis", guid)
+		// If the status filter is specified, don't deregister any registrations-without-resources
+		if filters.Statuses != nil && len(filters.Statuses) > 0 {
+			continue
 		}
-		servicesToDestroyByContainerIdBeforeDeregistration[containerId] = serviceObj
+
+		registrationsToDeregister[guid] = registration
 	}
 
-	// TODO PLEAAASE GO GENERICS... but we can't use 1.18 yet because it'll break all Kurtosis clients :(
-	var dockerOperation docker_operation_parallelizer.DockerOperation = func(
-		ctx context.Context,
-		dockerManager *docker_manager.DockerManager,
-		dockerObjectId string,
-	) error {
-		if err := dockerManager.RemoveContainer(ctx, dockerObjectId); err != nil {
-			return stacktrace.Propagate(err, "An error occurred removing user service container with ID '%v'", dockerObjectId)
-		}
-		return nil
-	}
-
-	successfulContainerRemoveGuidStrs, erroredContainerRemoveGuidStrs, err := docker_operation_parallelizer.RunDockerOperationInParallelForKurtosisObjects(
+	// Now try removing all the registrations-with-resources
+	successfulResourceRemovalGuids, erroredResourceRemovalGuids, err := backend.removeUserServiceDockerResources(
 		ctx,
-		servicesToDestroyByContainerIdBeforeDeregistration,
-		backend.dockerManager,
-		extractServiceGUIDFromServiceObj,
-		dockerOperation,
+		allServiceObjs,
+		allDockerResources,
 	)
 	if err != nil {
-		return nil, nil, stacktrace.Propagate(err, "An error occurred removing user service containers matching filters '%+v'", filters)
-	}
-
-	erroredGuids := map[service.ServiceGUID]error{}
-	for guidStr, err := range erroredContainerRemoveGuidStrs {
-		erroredGuids[service.ServiceGUID(guidStr)] = stacktrace.Propagate(
+		return nil, nil, stacktrace.Propagate(
 			err,
-			"An error occurred destroying container for service '%v'",
-			guidStr,
+			"An error occurred trying to remove user service Docker resources",
 		)
 	}
 
-	for guidStr := range successfulContainerRemoveGuidStrs {
-		guid := service.ServiceGUID(guidStr)
-		// Safe because earlier we verified that all the services that have containers also have GUIDs in the registration map
+	erroredGuids := map[service.ServiceGUID]error{}
+	for guid, err := range erroredResourceRemovalGuids {
+		erroredGuids[guid] = stacktrace.Propagate(
+			err,
+			"An error occurred destroying Docker resources for service '%v'",
+			guid,
+		)
+	}
+
+	for guid := range successfulResourceRemovalGuids {
 		registrationsToDeregister[guid] = matchingRegistrations[guid]
 	}
 
@@ -818,7 +876,9 @@ func (backend *DockerKurtosisBackend) getMatchingUserServiceDockerResources(
 	enclaveId enclave.EnclaveID,
 	maybeGuidsToMatch map[service.ServiceGUID]bool,
 ) (map[service.ServiceGUID]*userServiceDockerResources, error) {
-	// For the matching values, get the containers to check the status
+	result := map[service.ServiceGUID]*userServiceDockerResources{}
+
+	// Grab services, INDEPENDENT OF volumes
 	userServiceContainerSearchLabels := map[string]string{
 		label_key_consts.AppIDDockerLabelKey.GetString():         label_value_consts.AppIDDockerLabelValue.GetString(),
 		label_key_consts.EnclaveIDDockerLabelKey.GetString():     string(enclaveId),
@@ -828,7 +888,7 @@ func (backend *DockerKurtosisBackend) getMatchingUserServiceDockerResources(
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "An error occurred getting user service containers in enclave '%v' by labels: %+v", enclaveId, userServiceContainerSearchLabels)
 	}
-	resourcesByServiceGuid := map[service.ServiceGUID]*userServiceDockerResources{}
+
 	for _, container := range userServiceContainers {
 		serviceGuidStr, found := container.GetLabels()[label_key_consts.GUIDDockerLabelKey.GetString()]
 		if !found {
@@ -841,10 +901,48 @@ func (backend *DockerKurtosisBackend) getMatchingUserServiceDockerResources(
 				continue
 			}
 		}
-
-		resourcesByServiceGuid[serviceGuid] = &userServiceDockerResources{container: container}
+		
+		resourceObj, found := result[serviceGuid]
+		if !found {
+			resourceObj = &userServiceDockerResources{}
+		}
+		resourceObj.serviceContainer = container
+		result[serviceGuid] = resourceObj
 	}
-	return resourcesByServiceGuid, nil
+
+	// Grab volumes, INDEPENDENT OF whether there any containers
+	filesArtifactExpansionVolumeSearchLabels := map[string]string{
+		label_key_consts.AppIDDockerLabelKey.GetString():         label_value_consts.AppIDDockerLabelValue.GetString(),
+		label_key_consts.EnclaveIDDockerLabelKey.GetString():     string(enclaveId),
+		label_key_consts.VolumeTypeDockerLabelKey.GetString(): label_value_consts.FilesArtifactExpansionVolumeTypeDockerLabelValue.GetString(),
+	}
+	matchingFilesArtifactExpansionVolumes, err := backend.dockerManager.GetVolumesByLabels(ctx, filesArtifactExpansionVolumeSearchLabels)
+	if err != nil {
+		return nil, stacktrace.Propagate(err, "An error occurred getting files artifact expansion volumes in enclave '%v' by labels: %+v", enclaveId, filesArtifactExpansionVolumeSearchLabels)
+	}
+
+	for _, volume := range matchingFilesArtifactExpansionVolumes {
+		serviceGuidStr, found := volume.Labels[label_key_consts.UserServiceGUIDDockerLabelKey.GetString()]
+		if !found {
+			return nil, stacktrace.NewError("Found files artifact expansion volume '%v' that didn't have expected service GUID label '%v'", volume.Name, label_key_consts.UserServiceGUIDDockerLabelKey.GetString())
+		}
+		serviceGuid := service.ServiceGUID(serviceGuidStr)
+
+		if maybeGuidsToMatch != nil && len(maybeGuidsToMatch) > 0 {
+			if _, found := maybeGuidsToMatch[serviceGuid]; !found {
+				continue
+			}
+		}
+
+		resourceObj, found := result[serviceGuid]
+		if !found {
+			resourceObj = &userServiceDockerResources{}
+		}
+		resourceObj.expanderVolumeNames = append(resourceObj.expanderVolumeNames, volume.Name)
+		result[serviceGuid] = resourceObj
+	}
+
+	return result, nil
 }
 
 func getUserServiceObjsFromDockerResources(
@@ -852,8 +950,20 @@ func getUserServiceObjsFromDockerResources(
 	allDockerResources map[service.ServiceGUID]*userServiceDockerResources,
 ) (map[service.ServiceGUID]*service.Service, error) {
 	result := map[service.ServiceGUID]*service.Service{}
+
+	// If we have an entry in the map, it means there's at least one Docker resource
 	for serviceGuid, resources := range allDockerResources {
-		container := resources.container
+		container := resources.serviceContainer
+
+		// If we don't have a container, we don't have the service ID label which means we can't actually construct a Service object
+		// The only case where this would happen is if, during deletion, we delete the container but an error occurred deleting the volumes
+		if container == nil {
+			return nil, stacktrace.NewError(
+				"Service '%v' has Docker resources but not a container; this indicates that there the service's " +
+					"container was deleted but errors occurred deleting the rest of the resources",
+				serviceGuid,
+			)
+		}
 		containerName := container.GetName()
 		containerLabels := container.GetLabels()
 
@@ -1022,4 +1132,421 @@ func extractServiceGUIDFromServiceObj(uncastedObj interface{}) (string, error) {
 		return "", stacktrace.NewError("An error occurred downcasting the user service object")
 	}
 	return string(castedObj.GetRegistration().GetGUID()), nil
+}
+
+func (backend *DockerKurtosisBackend) doFilesArtifactExpansionAndGetUserServiceVolumes(
+	ctx context.Context,
+	serviceGuid service.ServiceGUID,
+	objAttrsProvider object_attributes_provider.DockerEnclaveObjectAttributesProvider,
+	freeIpAddrProvider *lib.FreeIpAddrTracker,
+	enclaveNetworkId string,
+	expanderImage string,
+	expanderEnvVars map[string]string,
+	expanderMountpointsToServiceMountpoints map[string]string,
+) (map[string]string, error) {
+	requestedExpanderMountpoints := map[string]bool{}
+	for expanderMountpoint := range expanderMountpointsToServiceMountpoints {
+		requestedExpanderMountpoints[expanderMountpoint] = true
+	}
+	expanderMountpointsToVolumeNames, err := backend.createFilesArtifactsExpansionVolumes(
+		ctx,
+		serviceGuid,
+		objAttrsProvider,
+		requestedExpanderMountpoints,
+	)
+	if err != nil {
+		return nil, stacktrace.Propagate(
+			err,
+			"Couldn't create files artifact expansion volumes for requested expander mounpoints: %+v",
+			requestedExpanderMountpoints,
+		)
+	}
+	shouldDeleteVolumes := true
+	defer func() {
+		if shouldDeleteVolumes {
+			for _, volumeName := range expanderMountpointsToVolumeNames {
+				// Use background context so we delete these even if input context was cancelled
+				if err := backend.dockerManager.RemoveVolume(context.Background(), volumeName); err != nil {
+					logrus.Errorf("Running the expansion failed so we tried to delete files artifact expansion volume '%v' that we created, but doing so threw an error:\n%v", volumeName, err)
+					logrus.Errorf("You'll need to delete volume '%v' manually!", volumeName)
+				}
+			}
+		}
+	}()
+
+	if err := backend.runFilesArtifactsExpander(
+		ctx,
+		serviceGuid,
+		objAttrsProvider,
+		freeIpAddrProvider,
+		expanderImage,
+		expanderEnvVars,
+		enclaveNetworkId,
+		expanderMountpointsToVolumeNames,
+	); err != nil {
+		return nil, stacktrace.Propagate(
+			err,
+			"An error occurred running files artifacts expander for service '%v'",
+			serviceGuid,
+		)
+	}
+
+	userServiceVolumeMounts := map[string]string{}
+	for expanderMountpoint, userServiceMountpoint := range expanderMountpointsToServiceMountpoints {
+		volumeName, found := expanderMountpointsToVolumeNames[expanderMountpoint]
+		if !found {
+			return nil, stacktrace.NewError(
+				"Found expander mountpoint '%v' for which no expansion volume was created; this should never happen " +
+					"and is a bug in Kurtosis",
+				expanderMountpoint,
+			)
+		}
+		userServiceVolumeMounts[volumeName] = userServiceMountpoint
+	}
+
+	shouldDeleteVolumes = false
+	return userServiceVolumeMounts, nil
+}
+
+// Runs a single expander container which expands one or more files artifacts into multiple volumes
+// NOTE: It is the caller's responsibility to handle the volumes that get returned
+func (backend *DockerKurtosisBackend) runFilesArtifactsExpander(
+	ctx context.Context,
+	serviceGuid service.ServiceGUID,
+	objAttrProvider object_attributes_provider.DockerEnclaveObjectAttributesProvider,
+	freeIpAddrProvider *lib.FreeIpAddrTracker,
+	image string,
+	envVars map[string]string,
+	enclaveNetworkId string,
+	mountpointsToVolumeNames map[string]string,
+) error {
+	containerAttrs, err := objAttrProvider.ForFilesArtifactsExpanderContainer(serviceGuid)
+	if err != nil {
+		return stacktrace.Propagate(err, "An error occurred while trying to get the files artifact expander container attributes for service '%v'", serviceGuid)
+	}
+	containerName := containerAttrs.GetName().GetString()
+	containerLabels := map[string]string{}
+	for labelKey, labelValue := range containerAttrs.GetLabels() {
+		containerLabels[labelKey.GetString()] = labelValue.GetString()
+	}
+
+	volumeMounts := map[string]string{}
+	for mountpointOnExpander, volumeName := range mountpointsToVolumeNames {
+		volumeMounts[volumeName] = mountpointOnExpander
+	}
+
+	ipAddr, err := freeIpAddrProvider.GetFreeIpAddr()
+	if err != nil {
+		return stacktrace.Propagate(err, "Couldn't get a free IP to give the expander container '%v'", containerName)
+	}
+	defer freeIpAddrProvider.ReleaseIpAddr(ipAddr)
+
+
+	createAndStartArgs := docker_manager.NewCreateAndStartContainerArgsBuilder(
+		image,
+		containerName,
+		enclaveNetworkId,
+	).WithStaticIP(
+		ipAddr,
+	).WithEnvironmentVariables(
+		envVars,
+	).WithVolumeMounts(
+		volumeMounts,
+	).WithLabels(
+		containerLabels,
+	).Build()
+	containerId, _, err := backend.dockerManager.CreateAndStartContainer(ctx, createAndStartArgs)
+	if err != nil {
+		return stacktrace.Propagate(
+			err,
+			"An error occurred creating files artifacts expander container '%v' for service '%v'",
+			containerName,
+			serviceGuid,
+		)
+	}
+	defer func() {
+		// We destroy the expander container, rather than leaving it around, so that we clean up the resource we created
+		// in this function (meaning the caller doesn't have to do it)
+		// We can do this because if an error occurs, we'll capture the logs of the container in the error we return
+		// to the user
+		if destroyContainerErr := backend.dockerManager.RemoveContainer(ctx, containerId); destroyContainerErr != nil {
+			logrus.Errorf(
+				"We tried to remove the expander container '%v' with ID '%v' that we started, but doing so threw an error:\n%v",
+				containerName,
+				containerId,
+				destroyContainerErr,
+			)
+			logrus.Errorf("ACTION REQUIRED: You'll need to remove files artifacts expander container '%v' manually", containerName)
+		}
+	}()
+
+	exitCode, err := backend.dockerManager.WaitForExit(ctx, containerId)
+	if err != nil {
+		return stacktrace.Propagate(
+			err,
+			"An error occurred waiting for files artifacts expander container '%v' to exit",
+			containerName,
+		)
+	}
+	if exitCode != expanderContainerSuccessExitCode {
+		containerLogsBlockStr, err := backend.getFilesArtifactsExpanderContainerLogsBlockStr(
+			ctx,
+			containerId,
+		)
+		if err != nil {
+			return stacktrace.NewError(
+				"Files artifacts expander container '%v' for service '%v' finished with non-%v exit code '%v' so we tried " +
+					"to get the logs, but doing so failed with an error:\n%v",
+				containerName,
+				serviceGuid,
+				expanderContainerSuccessExitCode,
+				exitCode,
+				err,
+			)
+		}
+		return stacktrace.NewError(
+			"Files artifacts expander container '%v' for service '%v' finished with non-%v exit code '%v' and logs:\n%v",
+			containerName,
+			serviceGuid,
+			expanderContainerSuccessExitCode,
+			exitCode,
+			containerLogsBlockStr,
+		)
+	}
+
+	return nil
+}
+
+// This seems like a lot of effort to go through to get the logs of a failed container, but easily seeing the reason an expander
+// container has failed has proven to be very useful
+func (backend *DockerKurtosisBackend) getFilesArtifactsExpanderContainerLogsBlockStr(
+	ctx context.Context,
+	containerId string,
+) (string, error) {
+	containerLogsReadCloser, err := backend.dockerManager.GetContainerLogs(ctx, containerId, shouldFollowContainerLogsWhenExpanderHasError)
+	if err != nil {
+		return "", stacktrace.Propagate(err, "An error occurred getting the logs for expander container with ID '%v'", containerId)
+	}
+	defer containerLogsReadCloser.Close()
+
+	buffer := &bytes.Buffer{}
+	concurrentBuffer := concurrent_writer.NewConcurrentWriter(buffer)
+
+	// TODO Push this down into GetContainerLogs!!!! This code actually has a bug where it won't work if the container is a TTY
+	//  container; the proper checking logic can be seen in the 'enclave dump' functions but should all be abstracted by GetContainerLogs
+	//  The only reason I'm not doing it right now is because we have the huge ETH deadline tomorrow and I don't have time for any
+	//  nice-to-have refactors (ktoday, 2022-05-22)
+	if _, err := stdcopy.StdCopy(concurrentBuffer, concurrentBuffer, containerLogsReadCloser); err != nil {
+		 return "", stacktrace.Propagate(
+			 err,
+			 "An error occurred copying logs to memory for files artifact expander container '%v'",
+			 containerId,
+		 )
+	}
+
+	wrappedContainerLogsStrBuilder := strings.Builder{}
+	wrappedContainerLogsStrBuilder.WriteString(fmt.Sprintf(
+		">>>>>>>>>>>>>>>>>>>>>>>>>>>>>> Logs for container '%v' <<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<\n",
+		containerId,
+	))
+	wrappedContainerLogsStrBuilder.WriteString(buffer.String())
+	wrappedContainerLogsStrBuilder.WriteString("\n")
+	wrappedContainerLogsStrBuilder.WriteString(fmt.Sprintf(
+		">>>>>>>>>>>>>>>>>>>>>>>>>>>> End logs for container '%v' <<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<<",
+		containerId,
+	))
+
+	return wrappedContainerLogsStrBuilder.String(), nil
+}
+
+// Takes in a list of mountpoints on the expander container that the expander container wants populated with volumes,
+// creates one volume per mountpoint location, and returns the volume_name -> mountpoint map that the container
+// can use when starting
+func (backend *DockerKurtosisBackend) createFilesArtifactsExpansionVolumes(
+	ctx context.Context,
+	serviceGuid service.ServiceGUID,
+	enclaveObjAttrsProvider object_attributes_provider.DockerEnclaveObjectAttributesProvider,
+	allMountpointsExpanderWants map[string]bool,
+) (map[string]string, error) {
+	shouldDeleteVolumes := true
+	result := map[string]string{}
+	for mountpointExpanderWants := range allMountpointsExpanderWants {
+		volumeAttrs, err := enclaveObjAttrsProvider.ForSingleFilesArtifactExpansionVolume(serviceGuid)
+		if err != nil {
+			return nil, stacktrace.Propagate(err, "An error occurred creating files artifact expansion volume for service '%v'", serviceGuid)
+		}
+		volumeNameStr := volumeAttrs.GetName().GetString()
+		volumeLabelsStrs := map[string]string{}
+		for key, value := range volumeAttrs.GetLabels() {
+			volumeLabelsStrs[key.GetString()] = value.GetString()
+		}
+		if err := backend.dockerManager.CreateVolume(
+			ctx,
+			volumeAttrs.GetName().GetString(),
+			volumeLabelsStrs,
+		); err != nil {
+			return nil, stacktrace.Propagate(
+				err,
+				"An error occurred creating files artifact expansion volume for service '%v' that's intended to be mounted " +
+					"on the expander container at path '%v'",
+				serviceGuid,
+				mountpointExpanderWants,
+			)
+		}
+		//goland:noinspection GoDeferInLoop
+		defer func() {
+			if shouldDeleteVolumes {
+				// Background context so we still run this even if the input context was cancelled
+				if err := backend.dockerManager.RemoveVolume(context.Background(), volumeNameStr); err != nil {
+					logrus.Warnf(
+						"Creating files artifact expansion volumes didn't complete successfully so we tried to delete volume '%v' that we created, but doing so threw an error:\n%v",
+						volumeNameStr,
+						err,
+					)
+					logrus.Warnf("You'll need to clean up volume '%v' manually!", volumeNameStr)
+				}
+			}
+		}()
+
+		result[mountpointExpanderWants] = volumeNameStr
+	}
+	shouldDeleteVolumes = false
+	return result, nil
+}
+
+/*
+Destroys the Docker containers and makes a best-effort attempt to destroy the volumes
+
+NOTE: We make a best-effort attempt to delete files artifact expansion volumes for a service, but there's still the
+possibility that some will get leaked! There's unfortunately no way around this though because:
+
+1) We delete the canonical resource (the user container) before the volumes out of necessity
+   since Docker won't let us delete volumes unless no containers are using them
+2) We can't undo container deletion
+3) Normally whichever resource is created first (volumes) is the one we'd use as the canonical resource,
+   but we can't do that since we're not guaranteed to have volumes (because a service may not have requested any
+   files artifacts).
+
+Therefore, we just make a best-effort attempt to clean up the volumes and leak the rest, though it's not THAT
+big of a deal since they'll be deleted when the enclave gets deleted.
+ */
+func (backend *DockerKurtosisBackend) removeUserServiceDockerResources(
+	ctx context.Context,
+	serviceObjectsToRemove map[service.ServiceGUID]*service.Service,
+	resourcesToRemove map[service.ServiceGUID]*userServiceDockerResources,
+) (map[service.ServiceGUID]bool, map[service.ServiceGUID]error, error) {
+
+	erroredGuids := map[service.ServiceGUID]error{}
+
+	// Before deleting anything, verify we have the same keys in both maps
+	if len(serviceObjectsToRemove) != len(resourcesToRemove) {
+		return nil, nil, stacktrace.NewError(
+			"Expected the number of services to remove (%v), and the number of services for which we have Docker resources for (%v), to be the same but were different!",
+			len(serviceObjectsToRemove),
+			len(resourcesToRemove),
+		)
+	}
+	for serviceGuid := range serviceObjectsToRemove {
+		if _, found := resourcesToRemove[serviceGuid]; !found {
+			return nil, nil, stacktrace.NewError(
+				"Have service object to remove '%v', which doesn't have corresponding Docker resources; this is a " +
+					"bug in Kurtosis",
+				serviceGuid,
+			)
+		}
+	}
+
+	uncastedKurtosisObjectsToRemoveByContainerId := map[string]interface{}{}
+	for serviceGuid, resources := range resourcesToRemove {
+		// Safe to skip the is-found check because we verified the map keys are identical earlier
+		serviceObj := serviceObjectsToRemove[serviceGuid]
+
+		containerId := resources.serviceContainer.GetId()
+		uncastedKurtosisObjectsToRemoveByContainerId[containerId] = serviceObj
+	}
+
+	// TODO Simplify this with Go generics
+	var dockerOperation docker_operation_parallelizer.DockerOperation = func(
+		ctx context.Context,
+		dockerManager *docker_manager.DockerManager,
+		dockerObjectId string,
+	) error {
+		if err := dockerManager.RemoveContainer(ctx, dockerObjectId); err != nil {
+			return stacktrace.Propagate(err, "An error occurred removing user service container with ID '%v'", dockerObjectId)
+		}
+		return nil
+	}
+
+	successfulContainerRemoveGuidStrs, erroredContainerRemoveGuidStrs, err := docker_operation_parallelizer.RunDockerOperationInParallelForKurtosisObjects(
+		ctx,
+		uncastedKurtosisObjectsToRemoveByContainerId,
+		backend.dockerManager,
+		extractServiceGUIDFromServiceObj,
+		dockerOperation,
+	)
+	if err != nil {
+		return nil, nil, stacktrace.Propagate(err, "An error occurred removing user service containers in parallel")
+	}
+
+	for guidStr, err := range erroredContainerRemoveGuidStrs {
+		erroredGuids[service.ServiceGUID(guidStr)] = stacktrace.Propagate(
+			err,
+			"An error occurred destroying container for service '%v'",
+			guidStr,
+		)
+	}
+
+	// TODO Parallelize if we need more perf (but we shouldn't, since removing volumes way faster than containers)
+	successfulVolumeRemovalGuids := map[service.ServiceGUID]bool{}
+	for serviceGuidStr := range successfulContainerRemoveGuidStrs {
+		serviceGuid := service.ServiceGUID(serviceGuidStr)
+
+		// Safe to skip the is-found check because we verified that the maps have the same keys earlier
+		resources := resourcesToRemove[serviceGuid]
+
+		failedVolumeErrStrs := []string{}
+		for _, volumeName := range resources.expanderVolumeNames {
+			/*
+			We try to delete as many volumes as we can here, rather than ejecting on failure, because any volumes not
+			deleted here will be leaked! There's unfortunately no way around this though because:
+
+			 1) we've already deleted the canonical resource (the user container) out of necessity
+			    since Docker won't let us delete volumes unless no containers are using them
+			 2) we can't undo container deletion
+			 3) normally whichever resource is created first (volumes) is the one we'd use as the canonical resource,
+			    but we can't do that since we're not guaranteed to have volumes
+
+			Therefore, we just make a best-effort attempt to clean up the volumes and leak the rest :(
+			 */
+			if err := backend.dockerManager.RemoveVolume(ctx, volumeName); err != nil {
+				errStrBuilder := strings.Builder{}
+				errStrBuilder.WriteString(fmt.Sprintf(
+					">>>>>>>>>>>>>>>>>> Removal error for volume %v <<<<<<<<<<<<<<<<<<<<<<<<<<<\n",
+					volumeName,
+				))
+				errStrBuilder.WriteString(err.Error())
+				errStrBuilder.WriteString("\n")
+				errStrBuilder.WriteString(fmt.Sprintf(
+					">>>>>>>>>>>>>>> End removal error for volume %v <<<<<<<<<<<<<<<<<<<<<<<<<<",
+					volumeName,
+				))
+				failedVolumeErrStrs = append(failedVolumeErrStrs, errStrBuilder.String())
+			}
+		}
+
+		if len(failedVolumeErrStrs) > 0 {
+			erroredGuids[serviceGuid] = stacktrace.NewError(
+				"Errors occurred removing volumes for service '%v'\n" +
+					 "ACTION REQUIRED: You will need to manually remove these volumes, else they will stay around until the enclave is destroyed!\n" +
+					 "%v",
+				serviceGuid,
+				strings.Join(failedVolumeErrStrs, "\n\n"),
+			)
+			continue
+		}
+		successfulVolumeRemovalGuids[serviceGuid] = true
+	}
+
+	successGuids := successfulVolumeRemovalGuids
+	return successGuids, erroredGuids, nil
 }